add confirm notification before deletion and fix vendor category issue

+++ resources/views/backend/content/edit.blade.php
@@ -110,7 +110,7 @@
 {{--                        </div>--}}
 
                         <div class="float-right">
-                            <a href="{{route("backend.content.destroy", $content)}}" class="btn btn-danger" data-method="DELETE" data-token="{{csrf_token()}}" data-toggle="tooltip" title="{{__('labels.backend.delete')}}"><i class="fas fa-trash-alt"></i></a>
+                            <a href="{{route("backend.content.destroy", $content)}}" class="btn btn-danger" data-method="DELETE" data-token="{{csrf_token()}}" data-toggle="tooltip" title="{{__('labels.backend.delete')}}" data-confirm="Are you sure?"><i class="fas fa-trash-alt"></i></a>
                             <a href="{{ route("backend.content.index") }}" class="btn btn-warning" data-toggle="tooltip" title="{{__('labels.backend.cancel')}}"><i class="fas fa-reply"></i> Cancel</a>
                         </div>
                     </div>
